requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-elasticsearch-dsl==7.1.4
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
drf-yasg==1.17.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
pika==1.1.0
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
